+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/query-management/components/query-event-completion/query-event-completion.component.spec.ts
@@ -1,6 +1,12 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { QueryEventCompletionComponent } from './query-event-completion.component';
+import { MockComponent } from 'ng2-mock-component';
+
+const queryEventCompletionComponentMock: any = MockComponent({
+  selector: 'ccd-case-event-completion',
+  inputs: ['eventCompletionParams']
+});
 
 describe('QueryEventCompletionComponent', () => {
   let component: QueryEventCompletionComponent;
@@ -8,7 +14,7 @@ describe('QueryEventCompletionComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ QueryEventCompletionComponent ]
+      declarations: [ QueryEventCompletionComponent, queryEventCompletionComponentMock ]
     })
     .compileComponents();
   });